+++ source/bf/chartable.h
@@ -10,15 +10,15 @@ extern void WriteCharMetrics(FILE *, indx);
 
 extern indx create_entry(char *name,
 			 char *filename,
-			 short width,
+			 int16_t width,
 			 boolean derived,
 			 indx dirix,
-			 short masters,
-			 short hintDir);
+			 int16_t masters,
+			 int16_t hintDir);
 
-extern long setnoncomps(void);
+extern int32_t setnoncomps(void);
 
-extern void AddCCtoTable(char *, BboxPtr, BboxPtr, short, indx);
+extern void AddCCtoTable(char *, BboxPtr, BboxPtr, int16_t, indx);
 
 extern void AddTransitionaltoTable(char *charname, Transitions *tr);
 
@@ -29,7 +29,7 @@ extern void FreeCharTab(char *);
 
 extern void sortchartab(boolean);
 
-extern void GetWidthandBbox(char *, short *, BboxPtr, boolean, indx);
+extern void GetWidthandBbox(char *, int16_t *, BboxPtr, boolean, indx);
 
 /* Returns whether the given character is in the char set and has
 a bez file. */
@@ -38,18 +38,18 @@ extern boolean CharExists(char *);
 /* Check that a bez file has been seen for the given file name */
 extern boolean CharFileExists(char *);
 
-extern void set_char_width(char *, long, short *, indx);
+extern void set_char_width(char *, int32_t, int16_t *, indx);
 
-extern long writechars(boolean, char **, long *, boolean, boolean);
+extern int32_t writechars(boolean, char **, int32_t *, boolean, boolean);
 
-extern long writesubrs(long *);
+extern int32_t writesubrs(int32_t *);
 
-extern long getCapDY(char *, boolean, indx);
+extern int32_t getCapDY(char *, boolean, indx);
 
 extern void writefontbbox(char *);
 
 extern void WriteBaseDesignBBox(FILE *, indx);
 
-extern long CharNameCost(void);
+extern int32_t CharNameCost(void);
 
 extern void SetCharEncoding(boolean, boolean);
